Move RowRange out of internal folder
